Testing Helper Functions
Wrote three test class stubs to test the functions that read/write
to the database and those that don't interact directly with the db.
Also made sure that the lat/lng fields on the register form are
required. Changed the model so the point for a station is not
nullable.