Started Maps Marker
Can get markers onto the page in the console. Need to figure out how
to get them to load with the page. Wrote get user by username test.
Need to complete test for closest bikes, need to get closest bikes
out of db in order to feed them to the map.