Shapely & Lat/Lng Methods
Wrote methods lat() and lng() on the Station class in model.py in
order to access the decimal version of the lat/lng values directly
from the station object instead of converting in the transaction
using ST_AsText. Will use this to pass the positions of the stations
directly through the template instead of sending an ajax request.
This will be better because now the google maps javascript doesn't
need to wait for the get requests to complete.