Update Home/Work Server Side
Added /update route to handle loading the update settings form and
updating that information in the db. Helper function
update_existing_user() takes request.form object and parses the
info for the db. update.html is currently empty but will need a
form with hidden fields like in the register form.